+++ test/rubocop/cop/minitest/literal_as_actual_argument_test.rb
@@ -99,4 +99,14 @@ class LiteralAsActualArgumentTest < Minitest::Test
       end
     RUBY
   end
+
+  def test_does_not_register_offense_when_expected_and_actual_are_basic_literals
+    assert_no_offenses(<<~RUBY)
+      class FooTest < Minitest::Test
+        def test_do_something
+          assert_equal 0, 1, 'message'
+        end
+      end
+    RUBY
+  end
 end